requirements.txt
streamlit
gtts
requests
duckduckgo-search==6.3.7
Pillow
